+++ js/main.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2022
+ * (c) Copyright Ascensio System SIA 2023
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -109,7 +109,7 @@
             winEditor.location.href = url;
         } else if (!OCA.Onlyoffice.setting.sameTab || OCA.Onlyoffice.mobile || OCA.Onlyoffice.Desktop) {
             winEditor = window.open(url, "_blank");
-        } else if ($("#isPublic").val() === "1" && !$("#filestable").length) {
+        } else if ($("#isPublic").val() === "1" && $("#mimetype").val() !== "httpd/unix-directory") {
             location.href = url;
         } else {
             var $iframe = $("<iframe id=\"onlyofficeFrame\" nonce=\"" + btoa(OC.requestToken) + "\" scrolling=\"no\" allowfullscreen src=\"" + url + "&inframe=true\" />");
@@ -477,7 +477,7 @@
     };
 
     OCA.Onlyoffice.GetFileExtension = function (fileName) {
-        var extension = fileName.substr(fileName.lastIndexOf(".") + 1).toLowerCase();
+        var extension = OCA.Onlyoffice.getFileExtension(fileName);
         return extension;
     };
 
@@ -493,7 +493,7 @@
     OCA.Onlyoffice.bindVersionClick = function () {
         OCA.Onlyoffice.unbindVersionClick();
         $(document).on("click.onlyoffice-version", "#versionsTabView .downloadVersion", function() {
-            var ext = $("#app-sidebar .fileName h3").text().split(".").pop();
+            var ext = OCA.Onlyoffice.GetFileExtension($("#app-sidebar .fileName h3").text());
             if (!OCA.Onlyoffice.setting.formats[ext]
                 || !OCA.Onlyoffice.setting.formats[ext].def) {
                 return true;
@@ -522,7 +522,7 @@
     }
 
     var initPage = function () {
-        if ($("#isPublic").val() === "1" && !$("#filestable").length) {
+        if ($("#isPublic").val() === "1" && $("#mimetype").val() !== "httpd/unix-directory") {
             var fileName = $("#filename").val();
             var extension = OCA.Onlyoffice.GetFileExtension(fileName);